+++ src/ts/EventHandlers/XPromoEndScreenEventHandler.ts
@@ -1,0 +1,17 @@
+import { EndScreenEventHandler, IEndScreenDownloadParameters } from 'EventHandlers/EndScreenEventHandler';
+import { XPromoCampaign } from 'Models/Campaigns/XPromoCampaign';
+import { IXPromoAdUnitParameters, XPromoAdUnit } from 'AdUnits/XPromoAdUnit';
+import { NativeBridge } from 'Native/NativeBridge';
+import { KeyCode } from 'Constants/Android/KeyCode';
+
+export class XPromoEndScreenEventHandler extends EndScreenEventHandler<XPromoCampaign, XPromoAdUnit> {
+    constructor(nativeBridge: NativeBridge, adUnit: XPromoAdUnit, parameters: IXPromoAdUnitParameters) {
+        super(nativeBridge, adUnit, parameters);
+    }
+
+    public onKeyEvent(keyCode: number): void {
+        if (keyCode === KeyCode.BACK && this._adUnit.isShowing() && !this._adUnit.isActive()) {
+            this._adUnit.hide();
+        }
+    }
+}