+++ src/ts/Parsers/ProgrammaticAdMobParser.ts
@@ -1,14 +1,42 @@
 import { CampaignParser } from 'Parsers/CampaignParser';
 import { Request } from 'Utilities/Request';
-import { Campaign } from 'Models/Campaign';
+import { Campaign, ICampaign } from 'Models/Campaign';
 import { NativeBridge } from 'Native/NativeBridge';
-import { AdMobCampaign } from 'Models/Campaigns/AdMobCampaign';
+import { AdMobCampaign, IAdMobCampaign } from 'Models/Campaigns/AdMobCampaign';
 import { AuctionResponse } from 'Models/AuctionResponse';
 import { Session } from 'Models/Session';
 
 export class ProgrammaticAdMobParser extends CampaignParser {
     public parse(nativeBridge: NativeBridge, request: Request, response: AuctionResponse, session: Session, gamerId: string, abGroup: number): Promise<Campaign> {
         const markup = response.getContent();
-        return Promise.resolve(new AdMobCampaign(markup, session, this.getProgrammaticCampaignId(nativeBridge), gamerId, abGroup, response.getCacheTTL(), response.getTrackingUrls(), response.getAdType(), response.getCreativeId(), response.getSeatId(), response.getCorrelationId()));
+        const cacheTTL = response.getCacheTTL();
+
+        const baseCampaignParams: ICampaign = {
+            id: this.getProgrammaticCampaignId(nativeBridge),
+            gamerId: gamerId,
+            abGroup: abGroup,
+            willExpireAt: cacheTTL ? Date.now() + cacheTTL * 1000 : undefined,
+            adType: response.getAdType() || undefined,
+            correlationId: response.getCorrelationId() || undefined,
+            creativeId: response.getCreativeId() || undefined,
+            seatId: response.getSeatId() || undefined,
+            meta: undefined,
+            appCategory: undefined,
+            appSubCategory: undefined,
+            advertiserDomain: undefined,
+            advertiserCampaignId: undefined,
+            advertiserBundleId: undefined,
+            useWebViewUserAgentForTracking: response.getUseWebViewUserAgentForTracking(),
+            buyerId: undefined,
+            session: session
+        };
+
+        const adMobCampaignParams: IAdMobCampaign = {
+            ... baseCampaignParams,
+            dynamicMarkup: markup,
+            tracking: response.getTrackingUrls() || undefined
+        };
+
+        return Promise.resolve(new AdMobCampaign(adMobCampaignParams));
     }
 }